feat(2981): address review comments from Max.
