+++ packages/backend/src/index.ts
@@ -307,6 +307,7 @@ export function initIocContainer(
       knex: await deps.use('knex'),
       accountingService: await deps.use('accountingService'),
       walletAddressService: await deps.use('walletAddressService'),
+      assetService: await deps.use('assetService'),
       config: await deps.use('config')
     })
   })
